more chef-15 updates
correctly do bundle install + rake install

didn't realize there was a baked in bundle install already, bring that
back.

use rake install instead of only relying on bundle install.  if the
chef gem we are trying to install needs to pave over an existing gem of
an identical version then bundler will not update it, so we have to
additionally rake install.  this is a detail which omnibus-software
definitions don't need to worry about because there's never a
pre-existing chef gem of the same version number containing different
code (which only happens in our system off on PRs forked from
master/current)

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>