Use rake install:local
We have all our deps installed from the bundle install so
use `rake install:local` to not do any rubygems depsolving.

For some reason this seems to fix the case when the main
branch has been advanced and we do not have a new major version
yet published to rubygems.

Which may be a rubygems bug?

At any rate it matches what we do in omnibus builds.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>